Remove unused function in utils.py and move `calculate_token_usage` to UsageProcessor
